examples/semantic-code-search/server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
huggingface-hub==0.16.4
optimum==1.16.1
chromadb==0.4.22
together==1.3.11
python-dotenv==1.0.0
gitpython==3.1.41
tree-sitter==0.20.2
numpy==1.26.3